atiqullah-ahmadzai/api/main/controllers/profile.py
"""
Dashboard
"""

from django.contrib.auth import authenticate
from django.views.decorators.csrf import csrf_exempt
from rest_framework.authtoken.models import Token
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny
from rest_framework.status import (
    HTTP_400_BAD_REQUEST,
    HTTP_404_NOT_FOUND,
    HTTP_200_OK
)
from rest_framework.response import Response
from django.contrib.sites.shortcuts import get_current_site
from urllib.parse import urlsplit, parse_qs
from werkzeug.datastructures import MultiDict
from django.shortcuts import redirect
from django.shortcuts import render
import os
from django.contrib.auth.decorators import login_required
from main.helpers.functions import *

#load login view
@csrf_exempt
@login_required
@permission_classes((AllowAny,))
def home(request):
    context = {
        'title': os.getenv('SITE_TITLE')+" - Dashboard",
    }
    return render(request, 'dashboard.html', context)