atiqullah-ahmadzai/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
filelock==3.15.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
libclang==18.1.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
poe-api-wrapper==1.4.2
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.3.4
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.10.1
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
termcolor==2.4.0
Theano==1.0.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
win32-setctime==1.1.0
wrapt==1.14.1
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
tensorboard==2.14.0
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
gensim==4.3.2
numpy==1.24.0
keras==2.14.0